rename feed updates to on demand
